# docs/openapi.contracts.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/audit-logs": {
      "get": {
        "operationId": "AuditController_list",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "501": {
            "description": "Contract only"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List club audit records",
        "tags": [
          "audit"
        ],
        "x-implementation-status": "contract-only"
      }
    },
    "/api/v1/audit-logs/{id}": {
      "get": {
        "operationId": "AuditController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "501": {
            "description": "Contract only"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get club audit record",
        "tags": [
          "audit"
        ],
        "x-implementation-status": "contract-only"
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "501": {
            "description": "Contract only"
          }
        },
        "summary": "Sign in with club account",
        "tags": [
          "auth"
        ],
        "x-implementation-status": "contract-only"
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "501": {
            "description": "Contract only"
          }
        },
        "summary": "Exchange refresh token",
        "tags": [
          "auth"
        ],
        "x-implementation-status": "contract-only"
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "501": {
            "description": "Contract only"
          }
        },
        "summary": "Revoke refresh token",
        "tags": [
          "auth"
        ],
        "x-implementation-status": "contract-only"
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "501": {
            "description": "Contract only"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current account and role",
        "tags": [
          "auth"
        ],
        "x-implementation-status": "contract-only"
      }
    },
    "/api/v1/auth/change-password": {
      "post": {
        "operationId": "AuthController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "501": {
            "description": "Contract only"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Change current account password",
        "tags": [
          "auth"
        ],
        "x-implementation-status": "contract-only"
      }
    },
...
    "/api/v1/horses/{horseId}/owners": {
      "get": {
        "operationId": "HorseDetailsController_owners",
        "parameters": [
          {
            "name": "horseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "501": {
            "description": "Contract only"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List horse ownership history",
        "tags": [
          "horses"
        ],
        "x-implementation-status": "contract-only"
      }
    },
    "/api/v1/horses/{horseId}/pedigree": {
      "get": {
        "operationId": "HorseDetailsController_pedigree",
        "parameters": [
          {
            "name": "horseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "501": {
            "description": "Contract only"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get horse pedigree",
        "tags": [
          "horses"
        ],
        "x-implementation-status": "contract-only"
      }
    },
    "/api/v1/horses/{horseId}/eligibility": {
      "get": {
        "operationId": "HorseDetailsController_eligibility",
        "parameters": [
          {
            "name": "horseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "501": {
            "description": "Contract only"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current training and racing eligibility",
        "tags": [
          "horses"
        ],
        "x-implementation-status": "contract-only"
      }
    },
    "/api/v1/horses/{horseId}/status": {
      "patch": {
        "operationId": "HorseDetailsController_status",
        "parameters": [
          {
            "name": "horseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateHorseStatusDto"
              }
            }
          }
        },
        "responses": {
          "501": {
            "description": "Contract only"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Change horse health or lifecycle status",
        "tags": [
          "horses"
        ],
        "x-implementation-status": "contract-only"
      }
    },
...
    "/api/v1/horses/{id}": {
      "get": {
        "operationId": "HorsesController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "501": {
            "description": "Contract only"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get horse profile",
        "tags": [
          "horses"
        ],
        "x-implementation-status": "contract-only"
      },
      "patch": {
        "operationId": "HorsesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateHorseDto"
              }
            }
          }
        },
        "responses": {
          "501": {
            "description": "Contract only"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update horse profile",
        "tags": [
          "horses"
        ],
        "x-implementation-status": "contract-only"
      },
      "delete": {
        "operationId": "HorsesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "501": {
            "description": "Contract only"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Soft-delete horse profile",
        "tags": [
          "horses"
        ],
        "x-implementation-status": "contract-only"
      }
    },
...
    "/api/v1/media/upload-requests": {
      "post": {
        "operationId": "MediaController_requestUpload",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestUploadDto"
              }
            }
          }
        },
        "responses": {
          "501": {
            "description": "Contract only"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Request time-limited private upload URL",
        "tags": [
          "media"
        ],
        "x-implementation-status": "contract-only"
      }
    },
    "/api/v1/media/{id}/complete": {
      "post": {
        "operationId": "MediaController_complete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "501": {
            "description": "Contract only"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Confirm completed object upload",
        "tags": [
          "media"
        ],
        "x-implementation-status": "contract-only"
      }
    },
    "/api/v1/media/{id}": {
      "get": {
        "operationId": "MediaController_media",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "501": {
            "description": "Contract only"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get media metadata",
        "tags": [
          "media"
        ],
        "x-implementation-status": "contract-only"
      }
    },
    "/api/v1/media/{id}/download-url": {
      "get": {
        "operationId": "MediaController_downloadUrl",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "501": {
            "description": "Contract only"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Request time-limited private download URL",
        "tags": [
          "media"
        ],
        "x-implementation-status": "contract-only"
      }
    },
...
  },
  "info": {
    "title": "Racehorse API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string"
          },
          "newPassword": {
            "type": "string",
            "minLength": 8
          }
        },
        "required": [
          "currentPassword",
          "newPassword"
        ]
      },
      "UpdateHorseStatusDto": {
        "type": "object",
        "properties": {
          "healthStatus": {
            "type": "string",
            "enum": [
              "ELIGIBLE",
              "UNDER_OBSERVATION",
              "INJURED",
              "QUARANTINED"
            ]
          },
          "lifecycleStatus": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "RETIRED",
              "TRANSFERRED"
            ]
          }
        }
      },
...
      "UpdateHorseDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "microchipId": {
            "type": "string"
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date"
          },
          "sireId": {
            "type": "string",
            "format": "uuid"
          },
          "damId": {
            "type": "string",
            "format": "uuid"
          },
          "photoAssetId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "RequestUploadDto": {
        "type": "object",
        "properties": {
          "fileName": {
            "type": "string"
          },
          "mimeType": {
            "type": "string"
          },
          "byteSize": {
            "type": "number",
            "minimum": 1
          }
        },
        "required": [
          "fileName",
          "mimeType",
          "byteSize"
        ]
      },
...
    }
  }
}